==> Veneto/Bim/5-LabelsPublishing.py <==
##
import acqua.labelCollection as lc
import os
import logging
path = '/Users/andrea/PycharmProjects/Acqua/Veneto/Bim/'
os.chdir(path)
geoJsonFile = 'Bim.geojson'
ll = lc.to_mongoDBInsertMany(geoJsonFile)
logging.info("Safe %s record(s) to MongoDB.",len(ll))
